+++ pybutton/request.py
@@ -20,7 +20,10 @@ if sys.version_info[0] == 3:
     from urllib.request import Request
     from urllib.request import urlopen
     from urllib.error import HTTPError
+    from urllib.parse import urlencode
     from urllib.parse import urlunsplit
+    from urllib.parse import urlparse
+    from urllib.parse import parse_qs
 
     def request(url, method, headers, data=None, timeout=None):
         ''' Make an HTTP request in Python 3.x
@@ -62,7 +65,10 @@ else:
     from urllib2 import Request
     from urllib2 import urlopen
     from urllib2 import HTTPError
+    from urllib import urlencode
     from urlparse import urlunsplit
+    from urlparse import urlparse
+    from urlparse import parse_qs
 
     def request(url, method, headers, data=None, timeout=None):
         ''' Make an HTTP request in Python 2.x
@@ -104,7 +110,7 @@ else:
             raise ButtonClientError('Invalid response: {0}'.format(response))
 
 
-def request_url(secure, hostname, port, path):
+def request_url(secure, hostname, port, path, query=None):
     '''
         Combines url components into a url passable into the request function.
 
@@ -113,13 +119,44 @@ def request_url(secure, hostname, port, path):
             hostname (str): The host name for the url.
             port (int): The port number, as an integer.
             path (str): The hierarchical path.
+            query (dict): A dict of query parameters.
 
         Returns:
             (str) A complete url made up of the arguments.
     '''
+    encoded_query = urlencode(query) if query else ''
     scheme = 'https' if secure else 'http'
     netloc = '{0}:{1}'.format(hostname, port)
 
-    return urlunsplit((scheme, netloc, path, '', ''))
+    return urlunsplit((scheme, netloc, path, encoded_query, ''))
 
-__all__ = [Request, urlopen, HTTPError, request, request_url]
+
+def query_dict(url):
+    '''
+        Given a url, returns a dictionary of its query parameters.
+
+        Args:
+            url (string): The url to extract query parameters from.
+
+        Returns:
+            (dict) A dictionary of query parameters, formatted as follows:
+            {
+              query_name: [ list of values ],
+              ...
+            }
+
+    '''
+    url_components = urlparse(url)
+
+    if (url_components):
+        query_string = url_components.query
+        return parse_qs(query_string)
+
+__all__ = [
+    Request,
+    urlopen,
+    HTTPError,
+    request,
+    request_url,
+    query_dict,
+]